Cleanup for GA
Update breaking Data fabric changes
