+++ 101-veracity-api/veracity-api-net/veracity-api-sample/Enums/AccessLevel.cs
@@ -1,0 +1,12 @@
+﻿namespace veracity_api_sample.Enums
+{
+    /// <summary>
+    /// Access level for container
+    /// </summary>
+    public enum AccessLevel
+    {
+        Owner = 2,
+        DataSteward = 1,
+        Consumer = 0
+    }
+}
