+++ 101-veracity-api/veracity-api-net/veracity-api-sample/Enums/KeyStatus.cs
@@ -1,0 +1,13 @@
+﻿namespace veracity_api_sample.Enums
+{
+    /// <summary>
+    /// Shows information about what type of keys are available for that resource
+    /// </summary>
+    public enum KeyStatus
+    {
+        NoKeys = 0,
+        Expired = 1,
+        Available = 2,
+        Active = 3
+    }
+}
