+++ 101-veracity-api/veracity-api-net/veracity-api-sample/Enums/StorageLocations.cs
@@ -1,0 +1,11 @@
+﻿namespace veracity_api_sample.Enums
+{
+    /// <summary>
+    /// Only those can be choosen when considering Storage Locations
+    /// </summary>
+    public enum StorageLocations
+    {
+        NorthEurope,
+        EastUs
+    }
+}
