+++ 101-veracity-api/veracity-api-net/veracity-api-sample/Models/Company.cs
@@ -1,0 +1,11 @@
+﻿namespace veracity_api_sample.Models
+{
+    /// <summary>
+    /// Json structure for Company
+    /// </summary>
+    public class Company
+    {
+        public string CompanyId { get; set; }
+        public string AzureId { get; set; }
+    }
+}
