+++ 101-veracity-api/veracity-api-net/veracity-api-sample/Models/Region.cs
@@ -1,0 +1,12 @@
+﻿namespace veracity_api_sample
+{
+    public class Region
+    {
+        public string ShortName { get; set; }
+        public string FullName { get; set; }
+        public string Location { get; set; }
+        public string AzureName { get; set; }
+        public string DisplayName { get; set; }
+        public string GroupName { get; set; }
+    }
+}
