+++ 101-veracity-api/veracity-api-net/veracity-api-sample/Models/ShareAccessResponse.cs
@@ -1,0 +1,10 @@
+﻿namespace veracity_api_sample.Models
+{
+    /// <summary>
+    /// Json structure for shared access response
+    /// </summary>
+    public class ShareAccessResponse
+    {
+        public string AccessSharingId { get; set; }
+    }
+}
