+++ 101-veracity-api/veracity-api-net/veracity-api-sample/Models/User.cs
@@ -1,0 +1,12 @@
+﻿namespace veracity_api_sample.Models
+{
+    /// <summary>
+    /// Json structure for User
+    /// </summary>
+    public class User
+    {
+        public string UserId { get; set; }
+        public string CompanyId { get; set; }
+        public string Role { get; set; }
+    }
+}
