+++ 101-veracity-api/veracity-api-net/veracity-api-sample/ProvisioningApi.cs
@@ -1,22 +1,13 @@
-﻿using System;
-using System.Net.Http;
+﻿using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
-using System.Web;
 using Newtonsoft.Json;
+using veracity_api_sample.Enums;
+using veracity_api_sample.Models;
 
 namespace veracity_api_sample
 {
-    /// <summary>
-    /// Only those can be choosen when considering Storage Locations
-    /// </summary>
-    public enum StorageLocations
-    {
-        Unknown,
-        NorthEurope,
-        EastUs1
-    }
     /// <summary>
     /// Sample class showing Veracity Provisioning API usage
     /// </summary>
@@ -24,6 +15,7 @@ namespace veracity_api_sample
     {
         private readonly string _baseProvisioningApiUrl;
         private readonly HttpClient _httpClient;
+
         /// <summary>
         /// Constructor taking bearerKey, subscriptionKey and baseUrl to provisioningAPI
         /// </summary>
@@ -37,6 +29,7 @@ namespace veracity_api_sample
             _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", bearerKey);
             _httpClient.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", subscriptionKey);
         }
+
         /// <summary>
         /// Creates new storage container in given location with optionally given name
         /// </summary>
@@ -45,16 +38,16 @@ namespace veracity_api_sample
         /// <returns>Error if operation failed, Accepted - when success</returns>
         public async Task<string> ProvisionContainer(StorageLocations storageLocation, string containerShortName = null)
         {
-            var queryString = HttpUtility.ParseQueryString(string.Empty);
-            if(!string.IsNullOrEmpty(containerShortName))
-                queryString["containerShortName"] = containerShortName;
+            var uri = $"{_baseProvisioningApiUrl}container";
 
-            var requestCode = Guid.NewGuid().ToString();
+            var containerInput = new ContainerInput
+            {
+                ContainerShortName = containerShortName,
+                MayContainPersonalData = false,
+                StorageLocation = storageLocation.ToString()
+            };
 
-            var uri = $"{_baseProvisioningApiUrl}container?storageLocation={storageLocation}&requestCode={requestCode}&{queryString}";
-            var body = JsonConvert.SerializeObject(new { StorageLocation = storageLocation.ToString(), RequestCode = requestCode, ContainerShortName = containerShortName });
-
-            var response = await _httpClient.PostAsync(uri, new StringContent(body, Encoding.UTF8, "application/json"));
+            var response = await _httpClient.PostAsync(uri, new StringContent(JsonConvert.SerializeObject(containerInput), Encoding.UTF8, "application/json"));
             var responseContent = await response.Content.ReadAsStringAsync();
             return response.IsSuccessStatusCode ? response.ReasonPhrase : responseContent;
         }
