+++ 101-veracity-api/veracity-api-net/veracity-api-sample/veracity-api-sample.csproj
@@ -50,15 +50,30 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Enums\AccessLevel.cs" />
+    <Compile Include="Enums\KeyStatus.cs" />
+    <Compile Include="Models\ContainerInput.cs" />
+    <Compile Include="Models\Region.cs" />
+    <Compile Include="Models\Access.cs" />
+    <Compile Include="Models\Accesses.cs" />
     <Compile Include="AdB2C.cs" />
+    <Compile Include="Models\Company.cs" />
     <Compile Include="DataApi.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ProvisioningApi.cs" />
+    <Compile Include="Models\Resource.cs" />
+    <Compile Include="Models\SasData.cs" />
+    <Compile Include="Models\ShareAccessResponse.cs" />
+    <Compile Include="Models\SharedResource.cs" />
+    <Compile Include="Models\StorageKeyTemplate.cs" />
+    <Compile Include="Enums\StorageLocations.cs" />
     <Compile Include="TokenCacheHelper.cs" />
+    <Compile Include="Models\User.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
